+++ src/mizdb_tomselect/static/mizdb_tomselect/js/popup_response.js
@@ -1,0 +1,3 @@
+/* global opener */
+const data = JSON.parse(document.getElementById('popup-response-data').dataset.popupResponse)
+opener.dismissPopup(window, data)