+++ tests/playwright_tests/test_dropdown_footer.py
@@ -1,11 +1,14 @@
+import re
+from urllib.parse import urlparse
+
 import pytest
 from django import forms
 from django.http import HttpResponse
 from django.urls import path, reverse
-from django.views.generic import FormView
+from django.views.generic import CreateView, FormView
 from playwright.sync_api import expect
 
-from mizdb_tomselect.views import AutocompleteView
+from mizdb_tomselect.views import AutocompleteView, PopupResponseMixin
 from mizdb_tomselect.widgets import MIZSelect
 from tests.testapp.models import Person
 
@@ -67,6 +70,13 @@ class NoFooterForm(forms.Form):
     )
 
 
+class PersonCreateView(PopupResponseMixin, CreateView):
+    template_name = "base.html"
+    model = Person
+    fields = ["full_name"]
+    success_url = "__SUCCESS_URL__"  # success url is irrelevant for popups
+
+
 urlpatterns = [
     path("autocomplete/", AutocompleteView.as_view(), name="autocomplete"),
     path("add_button/", FormView.as_view(form_class=AddForm, template_name="base.html"), name="add"),
@@ -75,10 +85,9 @@ urlpatterns = [
         "changelist_button/", FormView.as_view(form_class=ChangelistForm, template_name="base.html"), name="changelist"
     ),
     path("no_footer/", FormView.as_view(form_class=NoFooterForm, template_name="base.html"), name="no_footer"),
-    path("add/", lambda r: HttpResponse("This is a dummy add page."), name="add_page"),
+    path("add/", PersonCreateView.as_view(), name="add_page"),
     path("changelist/", lambda r: HttpResponse("This is a dummy changelist page."), name="changelist_page"),
 ]
-
 
 pytestmark = [pytest.mark.pw, pytest.mark.urls(__name__)]
 
@@ -109,9 +118,24 @@ def requests_add_page(get_url):
     """
 
     def inner(request):
-        return request.url == get_url("add_page")
+        return urlparse(request.url).path == urlparse(get_url("add_page")).path
 
     return inner
+
+
+@pytest.fixture
+def add_new_object(context, _page, add_button):
+    """
+    Create a new model object using the 'add' page.
+
+    Click the add button, fill out the form of the popup, and submit the form.
+    """
+    with context.expect_page() as new_page_info:
+        add_button.click()
+    new_page = new_page_info.value
+    new_page.wait_for_load_state()
+    new_page.get_by_label(re.compile("full name", re.IGNORECASE)).fill("Bob Testman")
+    new_page.get_by_role("button").click()
 
 
 @pytest.mark.parametrize(
@@ -196,6 +220,13 @@ class TestAddButtonClick:
         with _page.expect_popup(requests_add_page):
             add_button.click()
 
+    def test_add_updates_selection(self, _page, login_perms_user, add_new_object):
+        """
+        After adding a new object using the popup 'add' page, the created object
+        should be immediately selected.
+        """
+        expect(_page.locator(".ts-control .item")).to_have_text("Bob Testman")
+
 
 @pytest.mark.parametrize("view_name", ["create"])
 class TestCreateButtonClick:
@@ -204,7 +235,7 @@ class TestCreateButtonClick:
     def test_add_button_click_no_search_term(self, login_perms_user, _page, add_button, requests_add_page):
         """
         Assert that clicking the 'add' button with no search term opens the
-        'add' _page in a new tab.
+        'add' page in a new tab.
         """
         with _page.expect_popup(requests_add_page):
             add_button.click()
@@ -234,6 +265,24 @@ class TestCreateButtonClick:
             add_button.click()
         expect(_page.locator(".ts-control .item")).to_have_text("Bob Testman")
         expect(_page.locator(".dropdown-content")).not_to_be_visible()
+
+    def test_no_popup_with_search_term(self, context, login_perms_user, _page, add_button, search_input):
+        """
+        Assert that clicking the 'add' button with a search term does not open
+        the 'add' page.
+        """
+        with _page.expect_request_finished():
+            search_input.fill("Bob Testman")
+        with _page.expect_request_finished():
+            add_button.click()
+        assert len(context.pages) == 1
+
+    def test_add_updates_selection(self, _page, login_perms_user, add_new_object):
+        """
+        After adding a new object using the popup 'add' page, the created object
+        should be immediately selected.
+        """
+        expect(_page.locator(".ts-control .item")).to_have_text("Bob Testman")
 
 
 @pytest.mark.parametrize("view_name", ["changelist"])